requirements.txt
Flask
requests
pytz
python-dotenv
